Fashion Recommender/flashion_webapp/requirements.txt
Flask==1.1.2
gunicorn==19.6.0	
pandas>=0.20.3
numpy>=1.15.4
scikit-learn>=0.20.1
pillow>=6.2.2
imutils>=0.5.0
opencv-python-headless>=4.2.0
selenium>=3.141.0
keras>=2.3.0
tensorflow>=2.0.0b1